+++ dwds/lib/src/services/expression_compiler.dart
@@ -2,6 +2,21 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+/// Options passed to DDC and the expression compiler.
+class CompilerOptions {
+  final String moduleFormat;
+  final bool soundNullSafety;
+  final bool canaryFeatures;
+  final List<String> experiments;
+
+  CompilerOptions({
+    required this.moduleFormat,
+    required this.soundNullSafety,
+    required this.canaryFeatures,
+    required this.experiments,
+  });
+}
+
 /// Result of compilation of dart expression to JavaScript
 class ExpressionCompilationResult {
   final bool isError;
@@ -65,13 +80,10 @@ abstract class ExpressionCompiler {
   /// [updateDependencies] is called during isolate creation.
   Future<bool> updateDependencies(Map<String, ModuleInfo> modules);
 
-  /// Initializes the compiler with null safety mode and module format.
+  /// Initializes the compiler with the app build settings.
   ///
   /// May be called multiple times and always before [updateDependencies].
-  Future<void> initialize({
-    required String moduleFormat,
-    bool soundNullSafety = false,
-  });
+  Future<void> initialize(CompilerOptions options);
 }
 
 class ModuleInfo {