+++ dwds/test/events_test.dart
@@ -12,6 +12,7 @@ import 'package:test/test.dart';
 import 'package:test_common/logging.dart';
 import 'package:test_common/test_sdk_configuration.dart';
 import 'package:vm_service/vm_service.dart';
+import 'package:vm_service_interface/vm_service_interface.dart';
 import 'package:webdriver/async_core.dart';
 
 import 'fixtures/context.dart';
@@ -132,8 +133,8 @@ void main() {
           ),
         );
         await context.setUp(
+          testSettings: TestSettings(enableExpressionEvaluation: true),
           debugSettings: TestDebugSettings.withDevTools(context),
-          enableExpressionEvaluation: true,
         );
         vmService = context.debugConnection.vmService;
         keyboard = context.webDriver.driver.keyboard;