+++ dwds/test/frontend_server_breakpoint_test.dart
@@ -9,9 +9,11 @@ import 'package:test/test.dart';
 import 'package:test_common/logging.dart';
 import 'package:test_common/test_sdk_configuration.dart';
 import 'package:vm_service/vm_service.dart';
+import 'package:vm_service_interface/vm_service_interface.dart';
 
 import 'fixtures/context.dart';
 import 'fixtures/project.dart';
+import 'fixtures/utilities.dart';
 
 void main() {
   // Enable verbose logging for debugging.
@@ -33,8 +35,10 @@ void main() {
     setUpAll(() async {
       setCurrentLogWriter(debug: debug);
       await context.setUp(
-        compilationMode: CompilationMode.frontendServer,
-        verboseCompiler: verboseCompiler,
+        testSettings: TestSettings(
+          compilationMode: CompilationMode.frontendServer,
+          verboseCompiler: verboseCompiler,
+        ),
       );
     });